[#25905] Purge Expired Cache remove administrator/cache/index.html file
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@21413 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ libraries/joomla/cache/storage/file.php
@@ -197,6 +197,10 @@ class JCacheStorageFile extends JCacheStorage
 		// files older than lifeTime get deleted from cache
 		$files = $this->_filesInFolder($this->_root, '', true, true);
 		foreach($files As $file) {
+			if('index.html' == JFile::getName($file)) {
+				// Don't delete index.html files
+				continue;
+			}
 			$time = @filemtime($file);
 			if (($time + $this->_lifetime) < $this->_now || empty($time)) {
 				$result |= @unlink($file);
